implement authorization to resources using providers

+++ app/Http/Controllers/PostController.php
@@ -36,6 +36,7 @@ class PostController extends Controller
      */
     public function create()
     {
+        $this->authorize('create_post');
         $categories = Category::all();
 
         return view('create', compact('categories'));
@@ -46,6 +47,7 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create_post');
         $request->validate([
            'image' => ['image', 'required', 'max:2028'],
            'title' => ['required', 'max:255'],
@@ -80,6 +82,7 @@ class PostController extends Controller
      */
     public function edit(string $id)
     {
+        $this->authorize('edit_post');
         $post = Post::findOrFail($id);
         $categories = Category::all();
 
@@ -91,6 +94,7 @@ class PostController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('edit_post');
         $request->validate([
             'title' => ['required', 'max:255'],
             'category_id' => ['required', 'integer'],
@@ -123,6 +127,7 @@ class PostController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete_post');
         $post = Post::findOrFail($id);
         $post->delete();
 
@@ -130,12 +135,14 @@ class PostController extends Controller
     }
 
     public function trashed(){
+        $this->authorize('delete_post');
         $posts = Post::onlyTrashed()->paginate(5);
 
         return view('trashed', compact('posts'));
     }
 
     public function restore($id){
+        $this->authorize('delete_post');
         $post = Post::onlyTrashed()->findOrFail($id);
         $post->restore();
 
@@ -143,6 +150,7 @@ class PostController extends Controller
     }
 
     public function forceDelete($id){
+        $this->authorize('delete_post');
         $post = Post::onlyTrashed()->findOrFail($id);
         File::delete(public_path($post->image));
         $post->forceDelete();
